added A1_3

+++ CYBR505/Lab13.c
@@ -5,7 +5,7 @@ float AverageTwo(int num1, int num2);
 int Lab13()
 {
 	int score1, score2;
-	double ave;
+	float ave;
 	printf("Please enter the 2 numbers to average: ");
 	scanf_s("%d%d", &score1, &score2);
 	ave = AverageTwo(score1, score2);
@@ -19,6 +19,6 @@ int Lab13()
 // Outputs: 1 float
 float AverageTwo(int num1, int num2)
 {
-	float average = (num1 + num2) / 2.0;
+	float average = (num1 + num2) / (float)2;
 	return average;
 }
